chore: reduce code duplication
