+++ frontend/src/utils/helpers.ts
@@ -17,3 +17,32 @@ export const formatPersonNumber = (ssn: string) => {
     ? `${ssn.slice(0, ssn.length - 4)}-${ssn.slice(ssn.length - 4)}`
     : ssn;
 };
+
+// Convert Arabic-Indic (٠–٩) and Eastern Arabic-Indic (۰–۹) numerals to ASCII
+export function normalizeDigits(input: string): string {
+  if (!input) return input;
+  let out = '';
+  for (const ch of input) {
+    const code = ch.codePointAt(0);
+    if (code === undefined) continue;
+    if (code >= 0x0660 && code <= 0x0669) {
+      out += String.fromCodePoint(48 + (code - 0x0660));
+      continue;
+    }
+    if (code >= 0x06f0 && code <= 0x06f9) {
+      out += String.fromCodePoint(48 + (code - 0x06f0));
+      continue;
+    }
+    out += ch;
+  }
+  return out;
+}
+
+export function collapseSpaces(s: string): string {
+  return s.replaceAll(/[\s\u00A0]+/g, ' ').trim();
+}
+
+export function capitalizeWord(word: string): string {
+  const lower = word.toLocaleLowerCase('sv-SE');
+  return lower ? lower[0].toLocaleUpperCase('sv-SE') + lower.slice(1) : lower;
+}
